Move disjoint-set and default-map into a shared datastructures directory.
