+++ lsif/src/shared/datastructures/default-map.test.ts
