+++ lsif/src/shared/datastructures/default-map.ts
