+++ lsif/src/worker/importer/importer.ts
@@ -2,7 +2,7 @@ import { assertId, hashKey, mustGet, readEnvInt } from '../../shared/util'
 import { Correlator, ResultSetData, ResultSetId } from './correlator'
 import { createSqliteConnection } from '../../shared/database/sqlite'
 import { databaseInsertionDurationHistogram, databaseInsertionErrorsCounter } from '../metrics'
-import { DefaultMap } from '../../default-map'
+import { DefaultMap } from '../../shared/datastructures/default-map'
 import { Edge, MonikerKind, RangeId, Vertex } from 'lsif-protocol'
 import { EntityManager } from 'typeorm'
 import { gzipJSON } from '../../encoding'
